+++ package.json
@@ -18,7 +18,8 @@
     "react-resizable-rotatable-draggable": "^0.2.0",
     "react-rnd": "^10.4.1",
     "styled-reset": "^4.5.2",
-    "sweetalert2": "^11.10.2"
+    "sweetalert2": "^11.10.2",
+    "zustand": "^4.5.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.43",
