+++ src/GlobalStyles.js
@@ -1,8 +1,6 @@
 import { createGlobalStyle } from 'styled-components';
-import reset from 'styled-reset';
 
 const GlobalStyles = createGlobalStyle`
-  ${reset}
     a{
       text-decoration: none;
       color: inherit;
@@ -10,8 +8,7 @@ const GlobalStyles = createGlobalStyle`
     *{
         box-sizing: border-box;
     }
-    html, body, div, span, h1, h2, h3, h4, h5, h6, p, 
-    a, dl, dt, dd, ol, ul, li, form, label, table {
+    html{
         margin: 0;
         padding: 0;
         border: 0;
@@ -40,6 +37,7 @@ const GlobalStyles = createGlobalStyle`
     input:focus {
       outline: none;
     }
+    //여기까지
     // Break Point
     /* $tablet: 500px; */
     $desktop: 1024px;
