+++ src/components/CalendarPage/DateNotification.jsx
@@ -10,78 +10,56 @@ import { baseInstance } from '../../api/config';
 // 이미지 import
 import Cloud1 from '../../assets/img/Cloud1.png';
 import Cloud2 from '../../assets/img/Cloud2.png';
-import InnerImg1 from '../../assets/img/InnerImg/SelectInnerImg1.png';
-import InnerImg2 from '../../assets/img/InnerImg/SelectInnerImg2.png';
-import InnerImg3 from '../../assets/img/InnerImg/SelectInnerImg3.png';
-import InnerImg4 from '../../assets/img/InnerImg/SelectInnerImg4.png';
-import InnerImg5 from '../../assets/img/InnerImg/SelectInnerImg5.png';
-import InnerImg6 from '../../assets/img/InnerImg/SelectInnerImg6.png';
-import SelectImgBtn from '../../assets/img/SelectImgBtn.png';
-import MaskingTape1 from '../../assets/img/MaskingTape1.png';
-import MaskingTape2 from '../../assets/img/MaskingTape2.png';
-import BackBtn from '../../assets/img/BackBtn.png';
-import DiaryWritePen from '../../assets/img/DiaryWritePen.png';
-import RightNotificationImgLogo from '../../assets/img/RightNotificationImgLogo.png';
-import RightNotificationImgBubble from '../../assets/img/RightNotificationImgBubble.png';
+import { useDateNotificationStore } from '../store/useDateNotificationStore'; 
+import Case1 from './Case1';
+import Case2 from './Case2';
+import Case3 from './Case3';
 
+// DateNotification 컴포넌트 정의
 const DateNotification = ({
-  diarySettingPage,
-  setDiarySettingPage,
+  // diarySettingPage,
+  // setDiarySettingPage,
   diaryMonth,
   setDiaryMonth,
   diaryDay,
   setShareURL,
   shareURL,
 }) => {
-  const navigate = useNavigate();
+  // 기타 코드 생략
 
-  // const [diarySettingPage, setDiarySettingPage] = useState(2);
-  const [pageNum, setPageNum] = useState(1);
+  const { page, setPage } = useDateNotificationStore();
+  // const [InnerPageNum, setInnerPageNum] = useState(1);
+  const [diarySettingPage, setDiarySettingPage] = useState(null);
 
-  //날짜 임시 설정
-  const [diaryId, setDiaryId] = useState(null);
-  const [InnerPageNum, setInnerPageNum] = useState(1);
-
+  // 날짜 설정 페이지를 클릭했을 때의 동작을 처리하는 함수
+  // const handleWriteDiaryClick = () => {
+  //   navigate('../diary', { state: { innerPageNum: InnerPageNum } });
+  // };
 
   const diarySettingRef = useRef(null);
-  const maxInnerPaper = 6; //속지 종류 수
 
+ //diarySettingRef가 변경될 때마다 
   useEffect(() => {
     function handleClickOutside(event) {
-      if (
-        diarySettingRef.current &&
-        !diarySettingRef.current.contains(event.target)
-      ) {
+      // diarySettingRef가 존재하고 클릭된 요소가 ref 밖에 있다면 실행
+      if (diarySettingRef.current && !diarySettingRef.current.contains(event.target)) {
         setDiarySettingPage(1);
       }
     }
 
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [diarySettingRef]);
-
-  const handleWriteDiaryClick = () => {
-    navigate('../diary', { state: { innerPageNum: InnerPageNum } });
+  document.addEventListener('mousedown', handleClickOutside);
+  return () => {
+    document.removeEventListener('mousedown', handleClickOutside);
   };
-  
-  const PageNumSub = () => {
-    if (pageNum > 1) {
-      setPageNum((PageNum) => PageNum - 1); // 이전 상태를 가져와서 변경
-    }
-  };
+}, [diarySettingRef]);
 
-  const PageNumAdd = () => {
-    if (pageNum < maxInnerPaper) {
-      setPageNum((prevPageNum) => prevPageNum + 1); // 이전 상태를 가져와서 변경
-    }
-  };
 
+  // 클립보드에 텍스트 복사하는 함수
   const copyToClipboard = () => {
     navigator.clipboard
       .writeText(shareURL)
       .then(() => {
+        // 성공적으로 복사되었을 때의 알림
         Swal.fire({
           toast: true,
           position: 'top',
@@ -98,6 +76,7 @@ const DateNotification = ({
         });
       })
       .catch((error) => {
+        // 복사 실패 시의 알림
         console.error('클립보드 복사 실패:', error);
         Swal.fire({
           position: 'top-end',
@@ -114,117 +93,8 @@ const DateNotification = ({
       });
   };
 
-  const RotateImg = (pageNum) => {
-    switch (
-      pageNum // 이미지는 임시로 가져온거고 나중에 DB에서 가져와서 쓸 예정
-    ) {
-      case 1:
-        return <InnerImg src={InnerImg1} />;
-      case 2:
-        return <InnerImg src={InnerImg2} />;
-      case 3:
-        return <InnerImg src={InnerImg3} />;
-      case 4:
-        return <InnerImg src={InnerImg4} />;
-      case 5:
-        return <InnerImg src={InnerImg5} />;
-      case 6:
-        return <InnerImg src={InnerImg6} />;
-      default:
-        return null;
-    }
-  };
-
-  const createDiary = async () => {
-    try {
-      const response = await baseInstance.post(
-        '/diaries/',
-        {
-          day: `${diaryDay}`,
-          diary_bg_id: pageNum,
-        },
-      );
-      if ( response.status === 200) {
-        console.log('일기장 생성 성공');
-        setShareURL(response.data.sns_link); // sns_link를 state에 저장
-        setDiaryId(response.data.diary_id); //diary_id를 state에 저장
-        setInnerPageNum(response.data.diary_bg_id); // inner_page_num를 state에
-        setDiarySettingPage(3); // 페이지 변경
-    } else {
-      console.log('일기장 생성 실패');
-    }
-  } catch (error) {
-    console.error('API 호출 중 오류 발생 : ', error);
-  }
-};
- 
-  const renderDiarySettingPage = () => {
-    switch (diarySettingPage) {
-    case 1:
-    return (
-    
-          <DiarySettingWindow>
-            <DiarySettingImgLogo src={RightNotificationImgLogo} />
-            <DiarySettingImgBubble src={RightNotificationImgBubble} />
-            <DiarySettingText>
-              <div>친구들과 나의 특별한 순간을 작성할 날짜를 선택해주세요!</div>
-            </DiarySettingText>
-          </DiarySettingWindow>
-        );
-      case 2:
-        return (
-          <DiarySettingWindow>
-            <SelectDateText>
-              {diaryMonth}월 {diaryDay}일 일기를 작성해요!
-            </SelectDateText>
-            <SelectInnerPaperText>
-              일기 배경지를 선택해 주세요
-            </SelectInnerPaperText>
-
-            <TopMaskingTape src={MaskingTape1} />
-            {RotateImg(pageNum)}
-            <BottomMaskingTape src={MaskingTape2} />
-
-            <SelectImgLeftBtn src={SelectImgBtn} onClick={PageNumSub} />
-            <SelectImgRightBtn src={SelectImgBtn} onClick={PageNumAdd} />
-
-            <CheckBtn onClick={() => { createDiary()}}>확인</CheckBtn>
-          </DiarySettingWindow>
-        );
-      case 3:
-        return (
-          <DiarySettingWindow>
-            <BackButton src={BackBtn} onClick={() => setDiarySettingPage(2)} />
-
-            <SelectDateText2>
-              <div>
-                {diaryMonth}월 {diaryDay}일 일기를
-              </div>
-              <div>친구들에게 공유해 봐요!</div>
-            </SelectDateText2>
-
-            <ShareURL>{shareURL}</ShareURL>
-            <CopyBtn onClick={copyToClipboard}>복사</CopyBtn>
-            <Line />
-            <LetsWriteText>일기를 작성하러 가볼까요?</LetsWriteText>
-            <WriteDiaryBtn
-            onClick={handleWriteDiaryClick}>
-             {/* onClick={() => (window.location.href = `http://${shareURL}`)}> */}
-            {/* onClick={() =>  navigate('../diary')} */}
-              작성하기 <DiaryWritePenIcon src={DiaryWritePen} />
-            </WriteDiaryBtn>
-
-
-          </DiarySettingWindow>
-        );
-      default:
-        return <p>오류 코드 : '{diarySettingPage}'</p>;
-    }
-  };
-
-
   const InnerPaperRotate = () => {
-    switch (innerPageNum) {
+    switch (inpageNum) {
       case 1:
         return <InnerPaperImg src={MainInnerImg1} ref={diaryRef} />;
       case 2:
@@ -240,22 +110,40 @@ const DateNotification = ({
       default:
         return <InnerPaperImg src={MainInnerImg1} ref={diaryRef} />;
     }
-  };
-
+    };
+    
   return (
     <RightStickerContainer ref={diarySettingRef}>
       <StyledCloud1 src={Cloud1} alt="Cloud 1" />
       <StyledCloud2 src={Cloud2} alt="Cloud 2" />
-      {renderDiarySettingPage()}
+      {page === 1 && <Case1 
+      diaryMonth={diaryMonth}
+      diaryDay={diaryDay}
+      setShareURL={setShareURL}
+      />}
+
+      {page === 2 && <Case2 
+      diaryMonth={diaryMonth}
+      diaryDay={diaryDay}
+      setShareURL={setShareURL}
+      />}
+
+      {page === 3 && <Case3 
+      diaryMonth={diaryMonth}
+      diaryDay={diaryDay}
+      setShareURL={setShareURL}
+      />}
     </RightStickerContainer>
   );
 };
 
+export default DateNotification;
+
 const RightStickerContainer = styled.div`
-  width: 15.1875rem;
-  height: 41.0625rem;
-  border-radius: 1.875rem;
-  background: #e7eef9;
+  // width: 15.1875rem;
+  // height: 41.0625rem;
+  // border-radius: 1.875rem;
+  // background: #e7eef9;
 `;
 const StyledCloud1 = styled.img`
   position: absolute;
@@ -273,304 +161,3 @@ const StyledCloud2 = styled.img`
   left: -4.0625rem;
   z-index: 10;
 `;
-
-const DiarySettingWindow = styled.div`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-
-  z-index: 15;
-
-  color: #000;
-  font-family: Arial Black;
-  font-size: 1rem;
-  font-style: Bold;
-  font-weight: 900;
-  line-height: normal;
-`;
-
-const DiarySettingImgLogo = styled.img`
-  position: absolute;
-  width: 15.1875rem;
-  height: 15.80469rem;
-  flex-shrink: 0;
-  display: flex;
-  margin-top: 20rem;
-`;
-const DiarySettingImgBubble = styled.img`
-  position: absolute;
-  width: 14.25rem;
-  height: 13rem;
-  flex-shrink: 0;
-  display: flex;
-  margin-top: 7.56rem;
-`;
-
-const DiarySettingText = styled.div`
-  position: absolute;
-  width: 10rem;
-  color: #000;
-  font-family: 'mong';
-  font-size: 1.125rem;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 1.5rem;
-  display: flex;
-  margin-top: 11.12rem;
-`;
-
-const NotifyText = styled.div`
-  margin-top: 5rem;
-`;
-
-const SelectDateText = styled.p`
-  position: absolute;
-  margin-top: 4.69rem;
-  color: #2c199f;
-  font-family: Inter;
-  font-size: 1rem;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-  left: 1.19rem;
-`;
-
-const SelectInnerPaperText = styled.p`
-  position: absolute;
-  margin-top: 8.75rem;
-  color: #9f9f9f;
-  font-family: Inter;
-  font-size: 0.75rem;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-  left: 1.19rem;
-`;
-
-const InnerImg = styled.img`
-  position: absolute;
-  width: 12.3125rem;
-  height: 18.9375rem;
-  flex-shrink: 0;
-  margin-top: 11.12rem;
-  z-index: 2;
-`;
-
-const TopMaskingTape = styled.img`
-  position: absolute;
-  width: 3.75rem;
-  height: 1.04506rem;
-  transform: rotate(-30deg);
-  flex-shrink: 0;
-  margin-top: 11rem;
-  margin-left: -10rem;
-  z-index: 3;
-`;
-
-const BottomMaskingTape = styled.img`
-  position: absolute;
-  width: 3.75rem;
-  height: 1.04506rem;
-  transform: rotate(-30deg);
-  flex-shrink: 0;
-  margin-top: 29rem;
-  margin-right: -10rem;
-  z-index: 3;
-`;
-
-const SelectImgLeftBtn = styled.img`
-  position: absolute;
-  flex-shrink: 0;
-  height: 2.09863rem;
-  width: 1.59675rem;
-  margin-top: 31.06rem;
-  margin-left: -4rem;
-  z-index: 2;
-  cursor: pointer;
-`;
-
-const SelectImgRightBtn = styled.img`
-  position: absolute;
-  flex-shrink: 0;
-  height: 2.09863rem;
-  width: 1.59675rem;
-  flex-shrink: 0;
-  margin-top: 31.06rem;
-  margin-right: -4rem;
-  z-index: 2;
-  transform: rotate(-180deg);
-  cursor: pointer;
-`;
-
-const CheckBtn = styled.div`
-  position: absolute;
-  margin-top: 34.81rem;
-  width: 5.375rem;
-  height: 2.38881rem;
-  flex-shrink: 0;
-  border-radius: 1.25rem;
-  background: #c1c3ff;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  color: #fff;
-  font-family: Arial Black;
-  font-size: 1rem;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-  cursor: pointer;
-  transition: transform 0.3s ease-in-out;
-  &:hover {
-    transform: scale(1.1);
-  }
-`;
-const PageFrame = styled.div`
-  width: 15.1875rem;
-  height: 41.0625rem;
-  display: flex;
-  justify-content: center;
-
-  z-index: 15;
-
-  color: #000;
-  font-family: Arial Black;
-  font-size: 1rem;
-  font-style: Bold;
-  font-weight: 900;
-  line-height: normal;
-`;
-
-const SelectDateText2 = styled.p`
-  position: absolute;
-  margin-top: 6rem;
-  color: #2c199f;
-  font-family: Inter;
-  font-size: 1rem;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-  left: 1.19rem;
-`;
-
-const ShareURL = styled.div`
-  position: absolute;
-  margin-top: 11rem;
-  margin-right: 4rem;
-  width: 9rem;
-  height: 2rem;
-  flex-shrink: 0;
-  border-radius: 0.3125rem;
-  background: #fff;
-  display: flex;
-  align-items: center;
-  padding-left: 0.3rem;
-
-  color: #000;
-  font-family: Inter;
-  font-size: 0.75rem;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-  overflow: hidden;
-`;
-
-const CopyBtn = styled.button`
-  position: absolute;
-  margin-top: 11rem;
-  margin-left: 10rem;
-  width: 4rem;
-  height: 2rem;
-  flex-shrink: 0;
-  border-radius: 0.3125rem;
-  background: #cad9ff;
-
-  color: #fff;
-  font-family: Inter;
-  font-size: 0.75rem;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-
-  transition: transform 0.3s ease-in-out;
-  &:hover {
-    transform: scale(1.1);
-  }
-`;
-
-const Line = styled.hr`
-  position: absolute;
-  margin-top: 17.5rem;
-  width: 13.1875rem;
-  height: 0.0625rem;
-  background: #91b6ff;
-`;
-
-const LetsWriteText = styled.p`
-  position: absolute;
-  margin-top: 20.87rem;
-  width: 11.625rem;
-  color: #2c199f;
-  align-items: center;
-  justify-content: center;
-
-  font-family: Inter;
-  font-size: 1rem;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-`;
-
-const WriteDiaryBtn = styled.div`
-  position: absolute;
-  margin-top: 24.5rem;
-  width: 12.6875rem;
-  height: 2.38881rem;
-  flex-shrink: 0;
-  border-radius: 1.875rem;
-  background: #e7eef9;
-  border-radius: 1.25rem;
-  background: #c1c3ff;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  color: #fff;
-  font-family: Inter;
-  font-size: 1rem;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-  cursor: pointer;
-  transition: transform 0.3s ease-in-out;
-  &:hover {
-    transform: scale(1.1);
-  }
-`;
-
-const BackButton = styled.img`
-  position: absolute;
-  margin-top: 1rem;
-  margin-right: 12rem;
-  width: 1.625rem;
-  height: 1.625rem;
-  flex-shrink: 0;
-  cursor: pointer;
-  transition: transform 0.3s ease-in-out;
-  &:hover {
-    transform: scale(1.1);
-  }
-`;
-
-const DiaryWritePenIcon = styled.img`
-  position: absolute;
-  width: 1.625rem;
-  margin-bottom: 0.3rem;
-  margin-left: 6rem;
-  height: 1.625rem;
-  flex-shrink: 0;
-`;
-
-export default DateNotification;
